split DocumentEditor/blockquote code into shared/ui
