+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -33,7 +33,7 @@ import { withHeading } from './heading'
 import { nestList, unnestList, withList } from './lists'
 import { ComponentBlockContext, withComponentBlocks } from './component-blocks'
 import { getPlaceholderTextForPropPath } from './component-blocks/utils'
-import { withBlockquote } from './blockquote'
+import { withBlockquote } from './blockquote-shared'
 import { type Relationships, withRelationship } from './relationship'
 import { withDivider } from './divider'
 import { withCodeBlock } from './code-block'
